code/pubsub/split_json_file_per_schema/requirements.txt
google-cloud-storage
google-cloud-pubsub
json-schema-matcher

